Restrcuture record view tests [skip ci]
